+++ src/rpc/before_online.rs
@@ -1,10 +1,14 @@
+use crate::config::PicodataConfig;
+use crate::pgproto;
 use crate::tlog;
 use crate::traft::node;
 use crate::traft::{RaftIndex, RaftTerm};
+
 use std::time::Duration;
 
 crate::define_rpc_request! {
-    /// Enable all "enabled" plugins on instance locally.
+    /// Enables communication by PostgreSQL protocol and all
+    /// plugins on instance locally that are marked as "enabled".
     ///
     /// Called by governor at newly online instance.
     ///
@@ -13,11 +17,15 @@ crate::define_rpc_request! {
     /// 2. Storage failure
     /// 3. Timeout while waiting for an index from request
     /// 4. Request has an incorrect term - leader changed
-    /// 5. Any of "enabled" was loaded with errors
-    fn proc_enable_all_plugins(req: Request) -> crate::traft::Result<Response> {
+    /// 5. Address for pgproto cannot be parsed or is busy
+    /// 6. Any of "enabled" plugins was loaded with errors
+    fn proc_before_online(req: Request) -> crate::traft::Result<Response> {
         let node = node::global()?;
         node.wait_index(req.applied, req.timeout)?;
         node.status().check_term(req.term)?;
+
+        let instance_config = &PicodataConfig::get().instance;
+        pgproto::start(&instance_config.pg, instance_config.instance_dir(), &node.storage)?;
 
         let result = node.plugin_manager.handle_instance_online();
         if let Err(e) = result {